backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
pandas
langchain
langchain-openai
langchain-community
openai
